Tests: rename files that don't contain tests so they don't start with 'test'
