+++ test/TestConfig.py
@@ -16,7 +16,7 @@
 
 import unittest
 
-from test.TestTopydo import TopydoTest
+from test.TopydoTestCase import TopydoTest
 from topydo.lib.Config import config
