+++ test/TestDepCommand.py
@@ -16,7 +16,7 @@
 
 import unittest
 
-from test.TestCommand import CommandTest
+from test.CommandTestCase import CommandTest
 from topydo.commands.DepCommand import DepCommand
 from topydo.lib.TodoList import TodoList
