+++ test/TestFilter.py
@@ -21,7 +21,7 @@ from datetime import date, timedelta
 
 from test.Facilities import (load_file, load_file_to_todolist,
                              todolist_to_string)
-from test.TestTopydo import TopydoTest
+from test.TopydoTestCase import TopydoTest
 from topydo.lib import Filter
 from topydo.lib.Todo import Todo
