+++ test/TestListContextCommand.py
@@ -16,8 +16,8 @@
 
 import unittest
 
+from test.CommandTestCase import CommandTest
 from test.Facilities import load_file_to_todolist
-from test.TestCommand import CommandTest
 from topydo.commands.ListContextCommand import ListContextCommand
