+++ test/TestRecurrence.py
@@ -17,7 +17,7 @@
 import unittest
 from datetime import date, timedelta
 
-from test.TestTopydo import TopydoTest
+from test.TopydoTestCase import TopydoTest
 from topydo.lib.Config import config
 from topydo.lib.Recurrence import NoRecurrenceException, advance_recurring_todo
 from topydo.lib.Todo import Todo
