+++ test/TestTodoBase.py
@@ -20,7 +20,7 @@ import re
 import unittest
 from datetime import date, timedelta
 
-from test.TestTopydo import TopydoTest
+from test.TopydoTestCase import TopydoTest
 from topydo.lib.TodoBase import TodoBase
